fix: skip report-load call in widget preview
Preview widgets are served from the admin domain, where the
report-load POST hits a 307 redirect and ends in a 404, flooding
the api logs. Thread an isPreview flag through so the report-load
fetch is only emitted for real widgets.

Extracts getWidgetJavascriptOutput into a db-free module so it can
be unit tested in isolation.
